+++ project-base/src/Shopsys/ShopBundle/Controller/Front/ContactFormController.php
@@ -3,9 +3,11 @@
 namespace Shopsys\ShopBundle\Controller\Front;
 
 use Shopsys\ShopBundle\Component\Controller\FrontBaseController;
+use Shopsys\ShopBundle\Component\Domain\Domain;
 use Shopsys\ShopBundle\Form\Front\Contact\ContactFormType;
 use Shopsys\ShopBundle\Model\ContactForm\ContactFormData;
 use Shopsys\ShopBundle\Model\ContactForm\ContactFormFacade;
+use Shopsys\ShopBundle\Model\LegalConditions\LegalConditionsFacade;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -16,9 +18,24 @@ class ContactFormController extends FrontBaseController
      */
     private $contactFormFacade;
 
-    public function __construct(ContactFormFacade $contactFormFacade)
-    {
+    /**
+     * @var \Shopsys\ShopBundle\Model\LegalConditions\LegalConditionsFacade
+     */
+    private $legalConditionsFacade;
+
+    /**
+     * @var \Shopsys\ShopBundle\Component\Domain\Domain
+     */
+    private $domain;
+
+    public function __construct(
+        ContactFormFacade $contactFormFacade,
+        LegalConditionsFacade $legalConditionsFacade,
+        Domain $domain
+    ) {
         $this->contactFormFacade = $contactFormFacade;
+        $this->legalConditionsFacade = $legalConditionsFacade;
+        $this->domain = $domain;
     }
 
     /**
@@ -26,6 +43,8 @@ class ContactFormController extends FrontBaseController
      */
     public function sendAction(Request $request)
     {
+        $privacyPolicyArticle = $this->legalConditionsFacade->findPrivacyPolicy($this->domain->getId());
+
         $form = $this->createForm(ContactFormType::class, new ContactFormData(), [
             'action' => $this->generateUrl('front_contact_form_send'),
         ]);
@@ -48,6 +67,7 @@ class ContactFormController extends FrontBaseController
 
         $contactFormHtml = $this->renderView('@ShopsysShop/Front/Content/ContactForm/contactForm.html.twig', [
             'form' => $form->createView(),
+            'privacyPolicyArticle' => $privacyPolicyArticle,
         ]);
 
         return new JsonResponse([
@@ -58,12 +78,15 @@ class ContactFormController extends FrontBaseController
 
     public function indexAction()
     {
+        $privacyPolicyArticle = $this->legalConditionsFacade->findPrivacyPolicy($this->domain->getId());
+
         $form = $this->createForm(ContactFormType::class, new ContactFormData(), [
             'action' => $this->generateUrl('front_contact_form_send'),
         ]);
 
         return $this->render('@ShopsysShop/Front/Content/ContactForm/contactForm.html.twig', [
             'form' => $form->createView(),
+            'privacyPolicyArticle' => $privacyPolicyArticle,
         ]);
     }
 }